pyproject.toml
[tool.poetry]
name = "neural-network-pdes"
version = "0.1.0"
description = ""
authors = ["John <john.loverich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pytorch-lightning = "^1.6.3"
hydra-core = "^1.2.0"
matplotlib = "^3.5.2"
hydra-nevergrad-sweeper = "^1.2.0"
omegaconf = "^2.2.2"
protobuf = "<4.21.0"
torchvision = "^0.12.0"
Pillow = "^9.1.1"
torch = "^1.11.0"
functorch = "^0.1.1"
high-order-layers-torch = "^1.1.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
